[I25-152] refactor: InputOfModal 구조 개선 및 타입 단순화
- inputsOfModal 폴더를 inputs로 리네이밍
- FormFieldConfig 타입을 discriminated union으로 개선
- SkillTag와 Checkbox를 InputOfModal에 통합
- Config 구조 단순화 (타입별 필요한 필드만 정의)
- import 경로 정리 및 통일
- common.css 중복 제거
