+++ src/app/components/modal/inputs/LabelOfInputs.tsx
