+++ src/app/components/modal/inputs/common/common.css
@@ -1,5 +1,5 @@
 .input-common {
-  @apply w-full h-[3.3125rem] rounded-md cursor-pointer flex-center bg-light-gray-outline rounded-md;
+  @apply w-full h-[3.3125rem] rounded-md cursor-pointer flex-center bg-light-gray-outline;
 }
 
 .remove-input-focus {
