+++ src/app/components/modal/inputs/types/inputTypes.ts
@@ -1,7 +1,7 @@
 import { ChangeEvent } from 'react';
 import { InputConfig } from '../InputListProvider';
 
-export type InputType = 'text' | 'lock' | 'date' | 'edit' | 'search' | 'picture';
+export type InputType = 'text' | 'lock' | 'date' | 'edit' | 'search' | 'picture' | 'checkbox' | 'skillTag';
 
 // 기본 Input Props (picture 제외)
 export interface BaseInputPropsCommon {
@@ -32,13 +32,33 @@ export type LabelInputsProps = BaseInputProps & {
   label?: string;
 };
 
-// FormField: label과 InputListProvider 쌍의 설정
-export interface FormFieldConfig {
+// FormField 타입별 설정
+interface BaseFieldConfig {
   label: string;
   required?: boolean;
-  fieldName: string; // React Hook Form의 필드명
-  inputConfig: InputConfig; // InputListProvider의 설정
+  fieldName: string;
 }
+
+// InputList 타입
+interface InputListFieldConfig extends BaseFieldConfig {
+  type: 'inputList';
+  inputConfig: InputConfig;
+}
+
+// SkillTag 타입
+interface SkillTagFieldConfig extends BaseFieldConfig {
+  type: 'skillTag';
+  white?: boolean;
+}
+
+// Checkbox 타입
+interface CheckboxFieldConfig extends BaseFieldConfig {
+  type: 'checkbox';
+  label: string;
+  checkboxLabel?: string;
+}
+
+export type FormFieldConfig = InputListFieldConfig | SkillTagFieldConfig | CheckboxFieldConfig;
 
 // 전체 폼 설정
 export interface FormConfig {
